@@ -3,15 +3,12 @@ import 'package:flutter_modular/flutter_modular.dart';
 import 'package:ifbanews/app/app_widget.dart';
 import 'package:ifbanews/app/modules/home/home_module.dart';
 
-class AppModule extends MainModule {
+class AppModule extends Module {
   @override
-  List<Bind> get binds => [];
+  final List<Bind> binds = [];
 
   @override
-  List<Router> get routers => [
-        Router('/', module: HomeModule()),
-      ];
-
-  @override
-  Widget get bootstrap => AppWidget();
+  List<ModularRoute> routes = [
+    ModuleRoute('/', module: HomeModule()),
+  ];
 }
