@@ -10,14 +10,12 @@ class _AppWidgetState extends State<AppWidget> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      navigatorKey: Modular.navigatorKey,
       debugShowCheckedModeBanner: false,
       title: 'IFBA News',
       theme: ThemeData(
         primarySwatch: Colors.green,
       ),
-      onGenerateRoute: Modular.generateRoute,
       initialRoute: '/',
-    );
+    ).modular();
   }
 }
