@@ -5,6 +5,8 @@ import 'package:ifbanews/app/modules/home/home_controller.dart';
 import 'components/item_news.dart';
 
 class HomePage extends StatefulWidget {
+  const HomePage({Key? key}) : super(key: key);
+
   @override
   _HomePageState createState() => _HomePageState();
 }
@@ -22,32 +24,30 @@ class _HomePageState extends State<HomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("Notícias - IFBA Campus Irecê"),
+        title: const Text("Notícias - IFBA Campus Irecê"),
       ),
-      body: Container(
-        child: Observer(
-          builder: (context) {
-            if (controller.data != null) {
-              return ListView.builder(
-                itemCount: controller.data.length,
-                itemBuilder: (context, index) {
-                  final item = controller.data[index];
-                  return ItemNews(item);
-                },
-              );
-            } else if (controller.error) {
-              return Center(
-                child: Text(
-                  'Ocorreu um erro de conexão.',
-                  style: TextStyle(fontSize: 18.0),
-                ),
-              );
-            }
-            return Center(
-              child: CircularProgressIndicator(),
+      body: Observer(
+        builder: (context) {
+          if (controller.data != null) {
+            return ListView.builder(
+              itemCount: controller.data?.length,
+              itemBuilder: (context, index) {
+                final item = controller.data![index];
+                return ItemNews(item);
+              },
             );
-          },
-        ),
+          } else if (controller.error) {
+            return const Center(
+              child: Text(
+                'Ocorreu um erro de conexão.',
+                style: TextStyle(fontSize: 18.0),
+              ),
+            );
+          }
+          return const Center(
+            child: CircularProgressIndicator(),
+          );
+        },
       ),
     );
   }
